test: fix broken tests against Firebase auth
